refactor(api/src/ app&setHeaders): setHeader moved to middlewares
