+++ api/src/app.js
@@ -2,7 +2,7 @@ const express = require('express');
 const cookieParser = require('cookie-parser');
 const morgan = require('morgan');
 const cors = require('cors');
-const setHeaders =  require('./setHeaders');
+const setHeaders =  require('./middlewares/setHeaders');
 const routes = require('./routes/index');
 
 const server = express();
