+++ api/src/middlewares/setHeaders.js
